+++ src/Storefront/TwigExtension/ConfigLanguageBased.php
@@ -89,6 +89,7 @@ class ConfigLanguageBased extends AbstractExtension
         $salesChannelId = $this->getSalesChannelId($context);
 
         $parts = explode('.', $key);
+        /* @phpstan-ignore-next-line */
         if (empty($parts)) {
             return null;
         }
@@ -111,11 +112,7 @@ class ConfigLanguageBased extends AbstractExtension
         if (isset($context['context'])) {
             $salesChannelContext = $context['context'];
             if ($salesChannelContext instanceof SalesChannelContext) {
-                if ($salesChannelContext->getLanguageIdChain()) {
-                    if (count($salesChannelContext->getLanguageIdChain()) > 0) {
-                        return $salesChannelContext->getLanguageIdChain()[0];
-                    }
-                }
+                return $salesChannelContext->getLanguageIdChain()[0] ?? null;
             }
         }
         return null;